Add `lspsaga.nvim`
This plugin provides handy functionality about LSP actions.
This also allows for the removal of `goto-preview` and `code-action`
plugins.

Note: 'goto-definition' from LSPsaga will not be used since,  somehow it
alters the current buffer before jumping, which when done in the
'plugins.lua' file, makes Lazy think that a change was made and a
notification pops-up, which is irritating.