Add `mise` configuration with all the packages
The installation of those packages should now be as simple as:
`mise install` which will read the config.toml and install all the
packages. This commit also adds several other useful TUIs.

Will keep the installation methods from github for the other packages
for now for the sake of completeness.