Add `workspace-diagnostics` plugin
This allows for populating the diagnostics for all files in the
"workspace". It works by sending the `textDocument/didOpen` request to
all files in the project. As this can have performance impact in large
repos it will remain disabled by default. A keymapping is added to allow
populating the diagnostics on demand.